MODFORJ-9: Changed generated type of the decimal property from float into BigDecimal.

+++ modules/BusinessDomainDsl/org.mod4j.dsl.businessdomain.generator/src/main/java/org/mod4j/businessdomain/generator/helpers/BusinessClassHelpers.java
@@ -201,7 +201,7 @@ public class BusinessClassHelpers {
     }
 
     public static String javaNullableType(DecimalProperty p) {
-        return "Float";
+        return "BigDecimal";
     }
 
     public static String javaDefaultValue(DecimalProperty p) {
