+++ modules/Examples/RecordShop-ExampleProject/RecordShop-business/src/test/java/org/company/recordshop/business/data/RecordDaoTest.java
@@ -2,6 +2,8 @@ package org.company.recordshop.business.data;
 
 import static org.junit.Assert.assertEquals;
 
+import java.math.BigDecimal;
+
 import org.company.recordshop.business.recordshopcore.data.ArtistDao;
 import org.company.recordshop.business.recordshopcore.data.RecordDao;
 import org.company.recordshop.business.recordshopcore.domain.Artist;
@@ -28,7 +30,7 @@ public class RecordDaoTest extends AbstractDaoTestCase {
 
 	@Test
 	public void testRetrieve() {
-		Record record = new Record("Abbey Road", 25.50F, RecordTypeEnum.BLUERAY);
+		Record record = new Record("Abbey Road", BigDecimal.valueOf(25.50), RecordTypeEnum.BLUERAY);
 		recordDao.add(record);
 		flush();
 		clear();
@@ -42,7 +44,7 @@ public class RecordDaoTest extends AbstractDaoTestCase {
 		assertEquals(0, countRowsInTable("Record_TABLE"));
 		assertEquals(0, countRowsInTable("Artist_TABLE"));
 		assertEquals(0, countRowsInTable("records_contributors"));
-		Record record = new Record("Abbey Road", 25.50F, RecordTypeEnum.BLUERAY);
+		Record record = new Record("Abbey Road", BigDecimal.valueOf(25.50), RecordTypeEnum.BLUERAY);
 		Artist john = new Artist("John", "Lennon", date(), 
 				"singer/guitarist/songwriter");
 		Artist paul = new Artist("Paul", "McCartney", date(), "singer/bass/songwriter");
